refactor: reorgzine `composables/`
