+++ src/composables/usePopups.ts
@@ -4,26 +4,18 @@ import { getElement, getDoc } from '@/utils/commons'
 import { Logger, LoggerScopeDecorator } from '@/utils/logger'
 import { getArchiveLinkAnchor, getTorrentLinkAnchor } from '@/components/Gallery/utils/elements'
 
-export const torrentInnerHtml = ref<string>('')
-export const archiveInnerHtml = ref<string>('')
-export const favoriteInnerHtml = ref<string>('')
-
 /**
 * 預先載入 Torrent, Archive 和 Favorites 視窗
 * 同時把原先的 window.open() popup 改為在同一個頁面內的 popup
 */
 
-export function useFetchPopups() {
+export function usePopups() {
   const archiveLinkAnchor = getArchiveLinkAnchor()
   const torrentLinkAnchor = getTorrentLinkAnchor()
 
-  function getInnerHTMLs() {
-    return {
-      torrentInnerHtml,
-      archiveInnerHtml,
-      favoriteInnerHtml,
-    }
-  }
+  const torrentInnerHtml = ref<string>('')
+  const archiveInnerHtml = ref<string>('')
+  const favoriteInnerHtml = ref<string>('')
 
   async function preloadLinks() {
     [
@@ -147,7 +139,10 @@ export function useFetchPopups() {
   }
 
   return {
-    getInnerHTMLs,
+    torrentInnerHtml,
+    archiveInnerHtml,
+    favoriteInnerHtml,
+
     preloadLinks,
     fetchTorrents,
     fetchArchive,
